Refactored to ClaimsTable
